requirements.txt
pandas
amadeus
environ
python-dotenv
scikit-learn